split QueryBuilder into pieces and refactor
