+++ composer.json
@@ -18,7 +18,7 @@
     ],
     "require": {
       "php" : ">=5.4.0",
-      "michaels/data-manager": "dev-master",
+      "michaels/data-manager": "dev-develop",
       "ostico/phporient": "dev-master"
 
     },
