+++ tests/Unit/Drivers/OrientQueryProcessorTest.php
@@ -1,0 +1,18 @@
+<?php
+namespace Michaels\Spider\Test\Unit\Drivers;
+
+use Codeception\Specify;
+
+class OrientQueryBuilderTest extends \PHPUnit_Framework_TestCase
+{
+    use Specify;
+
+    public function testMethod()
+    {
+        $this->specify("it does something", function () {
+
+            $this->assertTrue(true);
+        });
+    }
+}
+
